fixed bugs and added functionality
- Started user menu design
- Ways of starting a game:
a. Hit enter when the game loads (This won't matter with new user menu
design.) (Focus is on the START button)
b. Hit space (Game.cs)
Then use the arrows to move.
- Game does end after dead (No movement allowed)
- Modified Checkdead method so it doesn't check after game is over
- Improved Two Players and Versus Computer Mode (Changed from checkboxes
to radio buttons)

NOTES:
- The location of the Modes radio buttons (2 players and Vs Computer)
and the START button in the Status bar is temporal, they will be placed
in a different window at the beginning of a new game.
- We need to improve the Two Players Mode. For example, the Checkdead
will be a little bit different in that mode.
- There are some other bugs and modifications I would like to discuss
with you when I meet you.